Fully automatic evaluation tuning and first phased eval tune(#121)
Test  | newAutogradImpl
Elo   | 12.83 +- 7.67 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.97 (-2.94, 2.94) [0.00, 5.00]
Games | N: 5664 W: 2281 L: 2072 D: 1311
Penta | [338, 484, 1069, 513, 428]

Bench: 1585157

+++ src/util/vec/avx2.hpp
@@ -9,7 +9,7 @@
 namespace Clockwork {
 
 forceinline u8 concat8(u8 a, u8 b) {
-    return a | (b << 4);
+    return static_cast<u8>(a | (b << 4));
 }
 forceinline u32 concat32(u16 a, u16 b) {
     return static_cast<u32>(a) | (static_cast<u32>(b) << 16);
